+++ test/OpenTelemetry.Exporter.OneCollector.Tests/OneCollectorLoggerProviderBuilderExtensionsTests.cs
@@ -22,6 +22,8 @@ public class OneCollectorLoggerProviderBuilderExtensionsTests
                 configure => configure.ConfigureBatchOptions(o => configurationInvocations++));
         });
 
+        _ = loggerFactory.CreateLogger("TestLogger");
+
         Assert.Equal(1, configurationInvocations);
     }
 
@@ -36,6 +38,8 @@ public class OneCollectorLoggerProviderBuilderExtensionsTests
                 "InstrumentationKey=token-extrainformation",
                 configure => configure.ConfigureExporter(exporter => exporterInstance = exporter));
         });
+
+        _ = loggerFactory.CreateLogger("TestLogger");
 
         Assert.NotNull(exporterInstance);
 
@@ -60,6 +64,8 @@ public class OneCollectorLoggerProviderBuilderExtensionsTests
                 configure => configure.ConfigureSerializationOptions(o => configurationInvocations++));
         });
 
+        _ = loggerFactory.CreateLogger("TestLogger");
+
         Assert.Equal(1, configurationInvocations);
     }
 
@@ -74,6 +80,8 @@ public class OneCollectorLoggerProviderBuilderExtensionsTests
                 "InstrumentationKey=token-extrainformation",
                 configure => configure.ConfigureTransportOptions(o => configurationInvocations++));
         });
+
+        _ = loggerFactory.CreateLogger("TestLogger");
 
         Assert.Equal(1, configurationInvocations);
     }
@@ -90,6 +98,8 @@ public class OneCollectorLoggerProviderBuilderExtensionsTests
                     configure => configure.SetConnectionString("InstrumentationKey=token-extrainformation"));
             },
             services => services.Configure<OneCollectorLogExporterOptions>(o => options = o));
+
+        _ = loggerFactory.CreateLogger("TestLogger");
 
         Assert.NotNull(options);
         Assert.Equal("InstrumentationKey=token-extrainformation", options.ConnectionString);
@@ -114,6 +124,8 @@ public class OneCollectorLoggerProviderBuilderExtensionsTests
             },
             services => services.Configure<OneCollectorLogExporterOptions>(o => options = o));
 
+        _ = loggerFactory.CreateLogger("TestLogger");
+
         Assert.NotNull(options);
         Assert.Equal(mappings, options.EventFullNameMappings);
     }
@@ -132,6 +144,8 @@ public class OneCollectorLoggerProviderBuilderExtensionsTests
             },
             services => services.Configure<OneCollectorLogExporterOptions>(o => options = o));
 
+        _ = loggerFactory.CreateLogger("TestLogger");
+
         Assert.NotNull(options);
         Assert.Equal("MyDefaultEventNamespace", options.DefaultEventNamespace);
     }
@@ -149,6 +163,8 @@ public class OneCollectorLoggerProviderBuilderExtensionsTests
                     configure => configure.SetDefaultEventName("MyDefaultEventName"));
             },
             services => services.Configure<OneCollectorLogExporterOptions>(o => options = o));
+
+        _ = loggerFactory.CreateLogger("TestLogger");
 
         Assert.NotNull(options);
         Assert.Equal("MyDefaultEventName", options.DefaultEventName);
@@ -179,6 +195,8 @@ public class OneCollectorLoggerProviderBuilderExtensionsTests
                 services.Configure<OneCollectorLogExporterOptions>(name, o => exporterOptions = o);
                 services.Configure<BatchExportLogRecordProcessorOptions>(name, o => batchOptions = o);
             });
+
+        _ = loggerFactory.CreateLogger("TestLogger");
 
         Assert.NotNull(exporterOptions);
         Assert.NotNull(batchOptions);
@@ -212,6 +230,7 @@ public class OneCollectorLoggerProviderBuilderExtensionsTests
             {
                 builder.AddOneCollectorExporter(configure => { });
             });
+            _ = loggerFactory.CreateLogger("TestLogger");
         });
 
         Assert.Throws<OneCollectorExporterValidationException>(() =>
@@ -220,6 +239,7 @@ public class OneCollectorLoggerProviderBuilderExtensionsTests
             {
                 builder.AddOneCollectorExporter("InstrumentationKey=invalidinstrumentationkey");
             });
+            _ = loggerFactory.CreateLogger("TestLogger");
         });
 
         Assert.Throws<OneCollectorExporterValidationException>(() =>
@@ -228,6 +248,7 @@ public class OneCollectorLoggerProviderBuilderExtensionsTests
             {
                 builder.AddOneCollectorExporter("UnknownKey=invalidinstrumentationkey");
             });
+            _ = loggerFactory.CreateLogger("TestLogger");
         });
     }
 
@@ -253,6 +274,8 @@ public class OneCollectorLoggerProviderBuilderExtensionsTests
                 services.Configure<BatchExportLogRecordProcessorOptions>(
                     o => configurationInvocations++);
             });
+
+        _ = loggerFactory.CreateLogger("TestLogger");
 
         Assert.Equal(2, configurationInvocations);
     }
@@ -286,14 +309,15 @@ public class OneCollectorLoggerProviderBuilderExtensionsTests
                     o => configurationInvocations++);
             });
 
+        _ = loggerFactory.CreateLogger("TestLogger");
+
         Assert.Equal(2, configurationInvocations);
     }
 
     private static ILoggerFactory CreateLoggerFactoryWithOneCollectorExporter(
         Action<LoggerProviderBuilder> configureLogging,
-        Action<IServiceCollection>? configureServices = null)
-    {
-        return LoggerFactory.Create(builder =>
+        Action<IServiceCollection>? configureServices = null) =>
+        LoggerFactory.Create(builder =>
         {
             builder.AddOpenTelemetry();
 
@@ -301,5 +325,4 @@ public class OneCollectorLoggerProviderBuilderExtensionsTests
 
             configureServices?.Invoke(builder.Services);
         });
-    }
 }