+++ src/OpenTelemetry.Instrumentation.EventCounters/CHANGELOG.md
@@ -6,6 +6,9 @@
   configured via `AddEventSources`.
   ([#4031](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4031))
 
+* Updated OpenTelemetry core component version(s) to `1.16.0`.
+  ([#4487](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4487))
+
 ## 1.15.1-alpha.1
 
 Released 2026-Apr-21